wishlist api created

+++ app/Http/Controllers/API/WishlistController.php
@@ -3,53 +3,69 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\WishListResource;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class WishlistController extends Controller
 {
     // get wishlists
     public function index()
     {
-        $wishlists = auth()->user()->wishlistItems()->with('product')->get();
-
-        return response()->json([
-            'data' => $wishlists,
-            'msg' => 'success',
-        ], 200);
+        $user = auth()->user();
+        if (!$user) {
+            return responseFail('User not found', 404);
+        }
+        return responseSuccess(WishListResource::collection($user->wishlistItems), 'Wishlist items', 200);
     }
 
     // add wishlist
 
     public function store(Request $request)
     {
-        $request->validate([
-            'product_id' => 'required|exists:products,id',
-        ]);
+        // return $request->all();
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'product_id' => 'required|exists:products,id',
+            ]
+        );
 
-        $wishlist = auth()->user()->wishlistItems()->create([
+        if ($validator->fails()) {
+            return response()->json(['status' => 422, 'message' => $validator->errors()], 422);
+        }
+
+        $user = $request->user();
+
+        if(!$user){
+            return responseFail('User not found', 404);
+        }
+        $wishlist = $user->wishlistItems()->where('product_id', $request->product_id)->first();
+        if ($wishlist) {
+            return responseFail('Product already in wishlist', 422);
+        }
+
+        $user->wishlistItems()->create([
             'product_id' => $request->product_id,
         ]);
 
-        return response()->json([
-            'data' => $wishlist,
-            'msg' => 'success',
-        ], 200);
+        
+        return responseSuccess(WishListResource::collection($user->wishlistItems),'Product added to wishlist', 200);
     }
 
     // remove wishlist
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $wishlist = auth()->user()->wishlistItems()->where('id', $id)->first();
+        $user = $request->user();
+        if (!$user) {
+            return responseFail('User not found', 404);
+        }
+        $wishlist = $user->wishlistItems()->where('id', $id)->first();
         if ($wishlist) {
             $wishlist->delete();
-            return response()->json([
-                'msg' => 'success',
-            ], 200);
+            return responseSuccess(WishListResource::collection($user->wishlistItems), 'Product removed from wishlist', 200);
         }
-        return response()->json([
-            'msg' => 'error',
-        ], 404);
-
+        return responseFail('Product not found in wishlist', 404);
     }
 }
